Changed UARTBASE from unsigned int to unsigned long int
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@934 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ lpc2119/conio.c
@@ -4,7 +4,7 @@
 /*                                                                            */
 /******************************************************************************/
 
-// $Id: conio.c,v 1.8 2007-10-29 16:12:18 cvs Exp $
+// $Id: conio.c,v 1.9 2008-01-12 21:21:15 cvs Exp $
 
 #include <conio.h>
 #include <cpu.h>
@@ -13,7 +13,7 @@
 
 /* Define relocatable UART register definitions */
 
-static unsigned int UARTBASE = UART0_BASE_ADDR;
+static unsigned long int UARTBASE = UART0_BASE_ADDR;
 
 #define UxRBR          (*(volatile unsigned long *)(UARTBASE + 0x00))
 #define UxTHR          (*(volatile unsigned long *)(UARTBASE + 0x00))
